[202412] New qsp port names on Arista 7060x6 platform (#1310)
#### Why I did it
To bring the port name aliases for the Arista-7060X6-64PE HWSKUs in line
with the SONiC port name convention doc:
https://github.com/sonic-net/SONiC/blob/1ff6da05e1a49ce0206df08653ca43969d1d36ca/doc/sonic-port-name.md

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it
Updated the port name aliases for the Arista-7060X6-64PE HWSKUs per the
SONiC convention doc.

#### How to verify it
Provision the target DUT. Once provisioned, execute `show interface
status` from the cli and confirm that the intended interface alias names
are present.

#### Which release branch to backport (provide reason below if selected)
- [x] 202412

Note: this is a cherry-pick to 202412 for original PR:
https://github.com/sonic-net/sonic-buildimage/pull/22955

#### Tested branch (Please provide the tested image version)
Tested on 202412.

#### Description for the changelog
Update the SONiC port name aliases for the Arista-7060X6-64PE HWSKUs.